Create interview hook
